requirements.txt
Flask
mysql-connector
connexion
connexion[swagger-ui]
pymongo
request
pprint
unittest
python-resize-image